@@ -5,7 +5,7 @@ homepage: https://github.com/OpenIoTHub
 publish_to: none
 
 environment:
-  sdk: ">=2.5.0 <3.0.0"
+  sdk: ^3.0.0
   flutter: ">=1.16.0"
 
 dependencies:
@@ -19,23 +19,23 @@ dependencies:
   flutter_natcloud_service:
 #    git: https://github.com/OpenIoTHub/flutter_natcloud_service
     path: ./flutter_natcloud_service/
-  http: ^0.12.0+2
+  http: ^0.13.0
   #  grpc
-  grpc: ^1.0.2
+  grpc: ^3.1.0
   #  grpc协议
-  protobuf: ^1.0.1
+  protobuf: ^2.0.0
   #  app内嵌网页浏览器
-  webview_flutter: ^0.3.19+9
+  webview_flutter: ^4.2.1
   #  打开url
-  url_launcher: ^5.0.2
+  url_launcher: ^6.1.11
   #  安卓intent调用其他程序
   android_intent: ^2.0.2
   #  数据存储
-  shared_preferences: ^0.5.2
+  shared_preferences: ^2.1.1
   #  事件总线
   event_bus: ^2.0.0
   #  动态请求权限，定位，存储等
-  permission_handler: ^3.1.0
+  permission_handler: ^10.2.0
   #  查看手机安装的app
 #  device_apps: ^1.0.8
   #  静态网页sever
@@ -43,21 +43,21 @@ dependencies:
   #  player
 #  video_player: ^0.10.1+3
   #  mDNS
-  multicast_dns: ^0.2.2
-  mdns_plugin: ^1.1.4
+  multicast_dns: ^0.3.2+3
+#  mdns_plugin: ^1.1.4
   #  sqflite: ^1.1.6+2
   #  webDAV的文件浏览器
 #  modules:
 #    git: https://github.com/OpenIoTHub/modules
-  openiothub_constants: ^0.0.1
-  openiothub_plugin: ^0.0.2+37
-  openiothub_api: ^0.0.2
-  openiothub_common_pages: ^0.0.2
+  openiothub_constants: ^1.0.0
+  openiothub_plugin: ^1.0.0
+  openiothub_api: ^1.0.0
+  openiothub_common_pages: ^1.0.0
 
-  openiothub_grpc_api: ^0.0.1
-  server_grpc_api: ^0.0.1
-  gateway_grpc_api: ^0.0.1
-  iot_manager_grpc_api: ^0.0.2
+  openiothub_grpc_api: ^1.0.0
+  server_grpc_api: ^1.0.0
+  gateway_grpc_api: ^1.0.0
+  iot_manager_grpc_api: ^1.0.0
 
 #    path: ../modules/
   # 应用OTA
