@@ -25,7 +25,7 @@ class CustomTheme with ChangeNotifier {
 
   Future<String> getThemeValue() async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
-    String themeValuePrefs;
+    String? themeValuePrefs;
     if (prefs != null) themeValuePrefs = prefs.getString("theme");
     return themeValuePrefs ?? themeValue;
   }
@@ -40,7 +40,10 @@ class CustomTheme with ChangeNotifier {
 
 class CustomThemes {
   static final Color _lightAccentColor = Colors.orange;
-  static final ThemeData light = ThemeData(brightness: Brightness.light,primarySwatch: _lightAccentColor, primaryColor: _lightAccentColor);
+  static final ThemeData light = ThemeData(
+      brightness: Brightness.light,
+      primarySwatch: _lightAccentColor as MaterialColor,
+      primaryColor: _lightAccentColor);
 
   static final ThemeData dark = ThemeData(brightness: Brightness.dark);
 }
