@@ -4,8 +4,6 @@ import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:openiothub/model/custom_theme.dart';
 import 'package:openiothub/util/ThemeUtils.dart';
-import 'package:openiothub_api/api/OpenIoTHub/CommonDeviceApi.dart';
-import 'package:openiothub_api/api/OpenIoTHub/SessionApi.dart';
 import 'package:openiothub_api/openiothub_api.dart';
 import 'package:openiothub_constants/constants/Constants.dart';
 import 'package:openiothub_grpc_api/pb/service.pb.dart';
@@ -15,7 +13,8 @@ import 'package:provider/provider.dart';
 import './commonDeviceServiceTypesList.dart';
 
 class CommonDeviceListPage extends StatefulWidget {
-  CommonDeviceListPage({Key key, this.title}) : super(key: key);
+  CommonDeviceListPage({required Key key, required this.title})
+      : super(key: key);
 
   final String title;
 
@@ -26,7 +25,7 @@ class CommonDeviceListPage extends StatefulWidget {
 class _CommonDeviceListPageState extends State<CommonDeviceListPage> {
   List<SessionConfig> _SessionList = [];
   List<Device> _CommonDeviceList = [];
-  Timer _timerPeriod;
+  late Timer _timerPeriod;
 
   @override
   void initState() {
@@ -53,8 +52,8 @@ class _CommonDeviceListPageState extends State<CommonDeviceListPage> {
         var listItemContent = ListTile(
           leading: Icon(Icons.devices,
               color: Provider.of<CustomTheme>(context).isLightTheme()
-                  ? CustomThemes.light.accentColor
-                  : CustomThemes.dark.accentColor),
+                  ? CustomThemes.light.primaryColorLight
+                  : CustomThemes.dark.primaryColorDark),
           title: Row(
             mainAxisAlignment: MainAxisAlignment.start,
             children: <Widget>[
@@ -184,6 +183,7 @@ class _CommonDeviceListPageState extends State<CommonDeviceListPage> {
           // 写成独立的组件，支持刷新
           return CommonDeviceServiceTypesList(
             device: device,
+            key: UniqueKey(),
           );
         },
       ),
@@ -238,7 +238,7 @@ class _CommonDeviceListPageState extends State<CommonDeviceListPage> {
                 Icon(Icons.cloud_done),
                 Expanded(
                     child: Text(
-                      "${pair.name}(${pair.description})",
+                  "${pair.name}(${pair.description})",
                   style: Constants.titleTextStyle,
                 )),
                 Constants.rightArrowIcon
