@@ -7,7 +7,8 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
+// ignore_for_file:unnecessary_string_interpolations, unnecessary_string_escapes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,18 +21,19 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'zh_CN';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "app_title" : MessageLookupByLibrary.simpleMessage("云易连"),
-    "profile_about_this_app" : MessageLookupByLibrary.simpleMessage("关于本软件"),
-    "profile_click_avatar_to_sign_in" : MessageLookupByLibrary.simpleMessage("点击头像登录"),
-    "profile_docs" : MessageLookupByLibrary.simpleMessage("文档"),
-    "profile_not_logged_in" : MessageLookupByLibrary.simpleMessage("未登录"),
-    "profile_servers" : MessageLookupByLibrary.simpleMessage("我的服务器"),
-    "profile_settings" : MessageLookupByLibrary.simpleMessage("设置"),
-    "profile_tools" : MessageLookupByLibrary.simpleMessage("工具"),
-    "tab_gateway" : MessageLookupByLibrary.simpleMessage("网关"),
-    "tab_host" : MessageLookupByLibrary.simpleMessage("主机"),
-    "tab_smart" : MessageLookupByLibrary.simpleMessage("智能"),
-    "tab_user" : MessageLookupByLibrary.simpleMessage("用户")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "app_title": MessageLookupByLibrary.simpleMessage("云亿连"),
+        "profile_about_this_app": MessageLookupByLibrary.simpleMessage("关于本软件"),
+        "profile_click_avatar_to_sign_in":
+            MessageLookupByLibrary.simpleMessage("点击头像登录"),
+        "profile_docs": MessageLookupByLibrary.simpleMessage("文档"),
+        "profile_not_logged_in": MessageLookupByLibrary.simpleMessage("未登录"),
+        "profile_servers": MessageLookupByLibrary.simpleMessage("我的服务器"),
+        "profile_settings": MessageLookupByLibrary.simpleMessage("设置"),
+        "profile_tools": MessageLookupByLibrary.simpleMessage("工具"),
+        "tab_gateway": MessageLookupByLibrary.simpleMessage("网关"),
+        "tab_host": MessageLookupByLibrary.simpleMessage("主机"),
+        "tab_smart": MessageLookupByLibrary.simpleMessage("智能"),
+        "tab_user": MessageLookupByLibrary.simpleMessage("用户")
+      };
 }
