@@ -1,22 +1,19 @@
-import 'dart:convert';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_natcloud_service/flutter_natcloud_service.dart';
-import 'package:fluttertoast/fluttertoast.dart';
 import 'package:jaguar/jaguar.dart';
 import 'package:jaguar_flutter_asset/jaguar_flutter_asset.dart';
 import 'package:openiothub_api/api/IoTManager/CnameManager.dart';
 import 'package:openiothub_api/api/OpenIoTHub/Utils.dart';
 import 'package:openiothub_constants/constants/Config.dart';
 import 'package:openiothub_constants/constants/WeChatConfig.dart';
+import 'package:openiothub_grpc_api/pb/service.pb.dart';
 // import 'package:shared_preferences/shared_preferences.dart';
 import 'package:wechat_kit/wechat_kit.dart';
 
-import 'package:openiothub_grpc_api/pb/service.pb.dart';
-
-Future<void> init(){
+Future<void> init() async {
   initBackgroundService();
   initHttpAssets();
   initWechat();
@@ -32,16 +29,16 @@ Future<void> initBackgroundService() async {
   });
 }
 
-Future<void> initHttpAssets(){
+Future<void> initHttpAssets() async {
   final server =
-  Jaguar(address: Config.webStaticIp, port: Config.webStaticPort);
+      Jaguar(address: Config.webStaticIp, port: Config.webStaticPort);
   server.addRoute(serveFlutterAssets());
   server.serve(logRequests: true).then((v) {
     server.log.onRecord.listen((r) => debugPrint("==serve-log：$r"));
   });
 }
 
-Future<void> initWechat(){
+Future<void> initWechat() async {
   // 初始化wechat_kit
   Wechat.instance.registerApp(
     appId: WeChatConfig.WECHAT_APPID,
@@ -49,15 +46,16 @@ Future<void> initWechat(){
   );
 }
 
-Future<void> initSystemUi(){
+Future<void> initSystemUi() async {
   //安卓透明状态栏
   if (Platform.isAndroid) {
-    SystemUiOverlayStyle systemUiOverlayStyle = SystemUiOverlayStyle(statusBarColor:Colors.transparent);
+    SystemUiOverlayStyle systemUiOverlayStyle =
+        SystemUiOverlayStyle(statusBarColor: Colors.transparent);
     SystemChrome.setSystemUIOverlayStyle(systemUiOverlayStyle);
   }
 }
 
-Future<void> loadConfig(){
+Future<void> loadConfig() async {
   Future.delayed(Duration(milliseconds: 500), () {
     UtilApi.SyncConfigWithToken().then((OpenIoTHubOperationResponse rsp) {
       // Fluttertoast.showToast(msg: rsp.msg);
