@@ -1,5 +1,5 @@
 import 'package:openiothub_api/api/OpenIoTHub/Utils.dart';
 
-Future<void> InitAllConfig() {
+Future<void> InitAllConfig() async {
   UtilApi.loadAllConfig();
 }
