@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 class Utf8Utils {
-  static String encode(String origin) {
+  static String? encode(String origin) {
     if (origin == null || origin.length == 0) {
       return null;
     }
@@ -14,13 +14,13 @@ class Utf8Utils {
     return result.substring(0, result.length - 1);
   }
 
-  static String decode(String encodeStr) {
+  static String? decode(String encodeStr) {
     if (encodeStr == null || encodeStr.length == 0) {
       return null;
     }
     List<String> list = encodeStr.split(",");
     if (list != null && list.isNotEmpty) {
-      List<int> intList = List();
+      List<int> intList = List<int>.empty();
       for (String s in list) {
         intList.add(int.parse(s));
       }
